% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_create_tidy_tables.R
\name{mbl_create_tidy_table}
\alias{mbl_create_tidy_table}
\title{Create tidy table with expression data from DGEList}
\usage{
mbl_create_tidy_table(y, log_transform = TRUE, prior.count = 1)
}
\arguments{
\item{y}{DGElist}

\item{log_transform}{Log tranform the CPMs?}

\item{prior.count}{Prior count to add to the CPMs}
}
\value{
a tidy data.frame
}
\description{
This function returns the log2(CPM + 1) values for all genes and samples
in the form of a tidy data.frame.
}
