% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_get_gene_annotation.R
\name{mbl_get_transcript_annotation}
\alias{mbl_get_transcript_annotation}
\title{Get transcript annotation file}
\usage{
mbl_get_transcript_annotation(organism = c("mouse", "fly", "fish",
  "planaria", "worm"))
}
\arguments{
\item{organism}{either "mouse", "fly", "planaria", "worm", or "fish"}
}
\value{
transcript level annotations
}
\description{
Get transcript annotation file
}
