% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_import_quant.R
\name{mbl_import_quantitation_results}
\alias{mbl_import_quantitation_results}
\title{Import salmon quantification results}
\usage{

  mbl_import_quantitation_results(path = path.expand("~/analysis/quantitation"),
  organism = c("mouse", "fly", "fish", "planaria", "worm"),
  rm.description = TRUE)
}
\arguments{
\item{path}{path to folder with quantification results}

\item{organism}{either "mouse", "fly", "planaria", "worm", or "fish"}
}
\value{
a DGElist object
}
\description{
Import salmon quantification results
}
