% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_load_data.R
\name{mbl_load_data}
\alias{mbl_load_data}
\title{Load DGEList with expression data from aws.s3}
\usage{
mbl_load_data(organism = c("mouse", "fly", "fish", "planaria", "worm"),
  dataset = c("pre_mbl", "mbl"))
}
\arguments{
\item{organism}{either "mouse", "fly", "planaria", "worm", or "fish"}

\item{dataset}{either "prem_mbl" or "mbl"}
}
\value{
a DGElist object
}
\description{
Load DGEList with expression data from aws.s3
}
