% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl_plot_pca.R
\name{mbl_plot_pca}
\alias{mbl_plot_pca}
\title{Plot the first principal compontents for a DGEList}
\usage{
mbl_plot_pca(x, intgroup = NA, prior.count = 5, ntop = 500)
}
\arguments{
\item{x}{DGEList}

\item{intgroup}{Character vector with column names of x$samples to color
the points}

\item{prior.count}{Number of pseudocounts to add before log transformation}

\item{ntop}{Number of most variable genes to use for PCA}
}
\value{
ggplot2 object
}
\description{
Plot the first principal compontents for a DGEList
}
\note{
This function was modified from the DESeq2::plotPCA function
}
