% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{samples_mbl}
\alias{samples_mbl}
\title{Sample annotations for MBL samples}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 213 rows and 12 columns.}
\usage{
samples_mbl
}
\description{
Sample annotations for MBL samples
}
\keyword{datasets}
