% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{samples_pre_mbl}
\alias{samples_pre_mbl}
\title{Sample annotations for pre-MBL samples}
\format{An object of class \code{data.frame} with 355 rows and 8 columns.}
\usage{
samples_pre_mbl
}
\description{
Sample annotations for pre-MBL samples
}
\keyword{datasets}
