% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\docType{data}
\name{samples_pre_mbl}
\alias{samples_pre_mbl}
\title{Sample annotations for pre-MBL samples}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 357 rows and 8 columns.}
\usage{
samples_pre_mbl
}
\description{
Sample annotations for pre-MBL samples
}
\keyword{datasets}
